refactor(VSparkline): remove bar variant
